US125506 refactor duplicate working copy code
