+++ components/d2l-activity-editor/d2l-activity-quiz-editor/state/quiz.js
@@ -1,98 +1,18 @@
 import { action, configure as configureMobx, decorate, observable } from 'mobx';
-import { fetchEntity } from '../../state/fetch-entity.js';
 import { QuizEntity } from 'siren-sdk/src/activities/quizzes/QuizEntity.js';
+import { WorkingCopy } from '../../state/working-copy.js';
 
 configureMobx({ enforceActions: 'observed' });
 
-export class Quiz {
+export class Quiz extends WorkingCopy {
 	constructor(href, token) {
+		super(Quiz, QuizEntity)
 		this.href = href;
 		this.token = token;
-		this._saving = null;
-		this._checkedOut = null;
-	}
-
-	async checkin(quizStore, refetch) {
-		if (!this._entity) {
-			return;
-		}
-
-		if (this._saving) {
-			return this._saving;
-		}
-
-		this._saving = this._entity.checkin();
-		let sirenEntity;
-		try {
-			sirenEntity = await this._saving;
-		} catch (e) {
-			return;
-		} finally {
-			this._saving = null;
-		}
-
-		if (!sirenEntity) return;
-		const href = sirenEntity.self();
-		const entity = new Quiz(href, this.token);
-		entity.load(sirenEntity);
-		quizStore.put(href, entity);
-
-		if (refetch) {
-			this.fetch(true);
-		}
-	}
-
-	checkout(quizStore, forcedCheckout) {
-		if (!forcedCheckout && this._checkedOut) {
-			return this._checkedOut;
-		}
-
-		let href = this.href;
-		const getHrefPromise = this._entity.checkout().then(sirenEntity => {
-			if (sirenEntity) {
-				href = sirenEntity.self();
-				const entity = new Quiz(href, this.token);
-				entity.load(sirenEntity);
-				quizStore.put(href, entity);
-			}
-			return href;
-		}, () => {
-			return href;
-		});
-
-		if (!forcedCheckout) {
-			this._checkedOut = getHrefPromise;
-		}
-
-		return getHrefPromise;
 	}
 
 	delete() {
 		return this._entity.delete();
-	}
-
-	async dirty(quizStore) {
-		const checkedOutHref = await this._checkedOut;
-		const checkedOutEntity = quizStore && quizStore.get(checkedOutHref);
-
-		// Check that this entity is not dirty, then check that it's checked out working copy does not have a `canCheckin` action.
-		// It avoids recursively fetching a working copy's working copy by not passing in a quizStore the second time.
-		const isQuizDirty = !this._entity.equals(this._makeQuizData()) || this._entity.canCheckin();
-		const isCheckedOutEntityDirty = checkedOutEntity && await checkedOutEntity.dirty();
-
-		return isQuizDirty || isCheckedOutEntityDirty;
-	}
-
-	async fetch(bypassCache) {
-		const sirenEntity = await fetchEntity(this.href, this.token, bypassCache);
-		if (sirenEntity) {
-			const entity = new QuizEntity(sirenEntity, this.token, {
-				remove: () => { },
-			});
-
-			this.load(entity);
-		}
-		return this;
 	}
 
 	load(entity) {
@@ -130,23 +50,6 @@ export class Quiz {
 		this.headerIsDisplayed = entity.headerIsDisplayed();
 		this.headerRichTextEditorConfig = entity.headerRichTextEditorConfig();
 		this.ipRestrictionsHref = entity.ipRestrictionsHref();
-	}
-
-	async save() {
-		if (!this._entity) {
-			return;
-		}
-
-		if (this._saving) {
-			return this._saving;
-		}
-
-		this._saving = this._entity.save(this._makeQuizData());
-
-		await this._saving;
-		this._saving = null;
-
-		await this.fetch();
 	}
 
 	setAutoSetGraded(isEnabled) {
@@ -193,7 +96,7 @@ export class Quiz {
 		this.isShuffleEnabled = isEnabled;
 	}
 
-	_makeQuizData() {
+	_makeEntityData() {
 		/* NOTE: if you add fields here, please make sure you update the corresponding equals method in siren-sdk.
 					 The cancel workflow is making use of that to detect changes.
 		*/
@@ -261,8 +164,5 @@ decorate(Quiz, {
 	setAutoSetGraded: action,
 	setDescription: action,
 	setHeader: action,
-	save: action,
-	delete: action,
-	checkout: action,
-	checkin: action
+	delete: action
 });
